Trekker ut komponentene som omfatter saksbehandlers input av meta-data ifbm opprettelse av en ny klagebehandling til en egen komponent
